requirements.txt
torch
torchvision
numpy
opencv-python
pillow
tqdm
pyyaml
scikit-image
